move Help/Doc over to overflowdb-traversal
