+++ console/src/main/scala/io/shiftleft/console/Help.scala
@@ -1,7 +1,6 @@
 package io.shiftleft.console
 
-import io.shiftleft.overflowdb.traversal.Doc
-import io.shiftleft.semanticcpg.utils.Table
+import io.shiftleft.overflowdb.traversal.help.{Doc, Table}
 import org.apache.commons.lang.WordUtils
 
 import scala.reflect.runtime.universe.{TypeTag, typeOf}
