+++ semanticcpg/build.sbt
@@ -4,11 +4,7 @@ dependsOn(Projects.codepropertygraph)
 
 libraryDependencies ++= Seq(
   "org.json4s" %% "json4s-native" % "3.6.7",
-  "com.massisframework" % "j-text-utils" % "0.3.4",
   "org.scala-lang.modules" %% "scala-collection-contrib" % "0.2.1",
-  "org.reflections" % "reflections" % "0.9.12",
-  "org.scala-lang" % "scala-reflect" % scalaVersion.value,
-  "org.scala-lang" % "scala-compiler" % scalaVersion.value,
   "org.scalatest" %% "scalatest" % "3.0.8" % Test,
   "io.shiftleft" %% "fuzzyc2cpg" % Versions.fuzzyc2cpg % Test
 )
