--- semanticcpg/src/main/scala/io/shiftleft/semanticcpg/Traversal.java
@@ -1,9 +1,0 @@
-package io.shiftleft.semanticcpg;
-
-import java.lang.annotation.*;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Traversal {
-  Class elementType();
-}
