+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/types/structure/Block.scala
@@ -2,12 +2,11 @@ package io.shiftleft.semanticcpg.language.types.structure
 
 import gremlin.scala.GremlinScala
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeTypes, nodes}
-import io.shiftleft.overflowdb.traversal.Doc
-import io.shiftleft.semanticcpg.Traversal
+import io.shiftleft.overflowdb.traversal.help.{Doc, TraversalExt}
 import io.shiftleft.semanticcpg.language.NodeSteps
 import io.shiftleft.semanticcpg.language._
 
-@Traversal(elementType = classOf[nodes.Block])
+@TraversalExt(elementType = classOf[nodes.Block])
 class Block(val wrapped: NodeSteps[nodes.Block]) extends AnyVal {
   private def raw: GremlinScala[nodes.Block] = wrapped.raw
